feat: Refactor object deletion to updated deletion rules
Since objects are no longer deleted directly, but only marked as 'trash'
and removed at a later time by a periodic mechanism, the process
leading up to deletion has changed significantly.

This is a verbose and readable implementation of the object deletion
with (hopefully) all its possible cases written out explicitly. This
can be used to implement an equally consistent but more performant
version.